Move hashing logic out of pipeline and into its own package.
Prelude to more invasive changes to the hashing interface.
