+++ hashing/BUILD
@@ -1,0 +1,39 @@
+cc_library (
+    name = "hashing",
+    srcs = [
+        "hashing.cc",
+    ],
+    hdrs = ["hashing.h"],
+    linkstatic = select({
+        "//tools/config:linkshared": 0,
+        "//conditions:default": 1,
+    }),
+    visibility = ["//visibility:public"],
+    deps = [
+        "//ast/treemap",
+        "//common/concurrency",
+        "//core",
+        "//main/options",
+        "//main/pipeline",
+    ]
+)
+
+cc_library (
+    name = "hashing-orig",
+    srcs = [
+        "hashing.cc",
+    ],
+    hdrs = ["hashing.h"],
+    linkstatic = select({
+        "//tools/config:linkshared": 0,
+        "//conditions:default": 1,
+    }),
+    visibility = ["//visibility:public"],
+    deps = [
+        "//ast/treemap",
+        "//common/concurrency",
+        "//core",
+        "//main/options",
+        "//main/pipeline:pipeline-orig",
+    ]
+)
