+++ main/pipeline/pipeline.h
@@ -46,11 +46,6 @@ typecheck(std::unique_ptr<core::GlobalState> &gs, std::vector<ast::ParsedFile> w
 
 ast::ParsedFile typecheckOne(core::Context ctx, ast::ParsedFile resolved, const options::Options &opts);
 
-// Computes file hashes for the given files, and stores them in the files. If supplied, attempts to retrieve hashes from
-// the key-value store. Returns 'true' if it had to compute any file hashes.
-void computeFileHashes(const std::vector<std::shared_ptr<core::File>> &files, spdlog::logger &logger,
-                       WorkerPool &workers);
-
 core::StrictLevel decideStrictLevel(const core::GlobalState &gs, const core::FileRef file,
                                     const options::Options &opts);
